requirements.txt
aiohttp==3.7.3
aniso8601==7.0.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.1.5
attrs==19.3.0
autopep8==1.5.4
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.2.1
blinker==1.4
boto3==1.16.20
botocore==1.19.20
Brlapi==0.7.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
cliapp==1.20180812.1
Click==7.0
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
dicttoxml==1.7.4
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap4==0.0.8
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-decouple==2.1
django-filter==2.4.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
GDAL==3.2.1
geocoder==1.38.1
geographiclib==1.50
geoip2==4.1.0
geopy==2.0.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
ip2geotools==0.1.5
IP2Location==8.5.1
iso3166==1.0.1
isort==5.6.4
itypes==1.2.0
jdcal==1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.0
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.5.2
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
mapbox==0.18.0
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.1.2
maxminddb==2.0.3
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.0
multidict==5.1.0
netifaces==0.10.4
numexpr==2.7.1
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
packaging==20.3
pandas==0.25.3
paramiko==2.6.0
patsy==0.5.1
pexpect==4.6.0
Pillow==7.0.0
pip-review==1.1.0
pkginfo==1.6.1
pluggy==0.13.0
polyline==1.4.0
promise==2.3
protobuf==3.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.8.1
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyquery==1.4.3
pyRFC3339==1.1
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.3
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
ratelim==0.1.6
readme-renderer==28.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.4.0
Rx==1.6.1
s3transfer==0.3.3
scikit-learn==0.24.1
scipy==1.3.3
seaborn==0.10.0
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.16.0
singledispatch==3.6.2
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==1.9.5
sqlparse==0.4.1
stripe==2.56.0
systemd-python==234
tables==3.6.1
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.54.1
ttystatus==0.38
twine==3.2.0
typed-ast==1.4.1
typing==3.7.4.3
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
whitenoise==4.1.2
wrapt==1.12.1
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
yarl==1.6.3
zipp==1.0.0
